[RSDK-8388] remove dead code (#15)
This is the super-easy, low-hanging fruit. There might be more, but it requires reading through 2,000 lines of mostly-live code.

Also, if you have a Makefile, people will expect that that's how to build stuff (no one will think the makefile is for installing optional tools, and building the module is done manually). So, I added a default target in the makefile to build the module itself. I've tried this both in a canon shell and in my laptop's normal environment, and can successfully generate `bin/Linux-aarch64/pinctrl` and `bin/Linux-x86_64/pinctrl`, respectively.


* add a default Makefile target

* remove unused I2C code

* go mod tidy

* remove unused SPI code